Move the tests to a sub-namespace
