+++ composer.json
@@ -8,6 +8,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^7.0 || ^8.0",
+        "rector/rector": "^0.5.5",
         "squizlabs/php_codesniffer": "^3.0"
     },
     "autoload": {
@@ -17,7 +18,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "pxgamer\\Mnemonics\\": "tests"
+            "pxgamer\\Mnemonics\\Tests\\": "tests"
         }
     },
     "scripts": {
